+++ Maleev_V_A_ISEbd21/Form1.cs
@@ -13,7 +13,7 @@ namespace Maleev_V_A_ISEbd21
     public partial class Form1 : Form
     {
 
-        private Benzovoz car;
+        private Itest car;
 
 
         public Form1()
@@ -36,8 +36,7 @@ namespace Maleev_V_A_ISEbd21
         private void buttonCreate_Click(object sender, EventArgs e)
         {
             Random rnd = new Random();
-            car = new Benzovoz(rnd.Next(100, 300), rnd.Next(1000, 2000), Color.Blue,
-           Color.Yellow);
+            car = new Truck(rnd.Next(100, 300), rnd.Next(1000, 2000), Color.Blue);
             car.SetPosition(rnd.Next(10, 100), rnd.Next(10, 100), pictureBoxCars.Width,
            pictureBoxCars.Height);
             Draw();
@@ -67,6 +66,15 @@ namespace Maleev_V_A_ISEbd21
                     break;
             }
             Draw();
-        }
+        }
+
+        private void buttonCreateBenz_Click(object sender, EventArgs e)
+        {
+            Random rnd = new Random();
+            car = new Benzovoz(rnd.Next(100, 300), rnd.Next(1000, 2000), Color.Gray, Color.Orange);
+            car.SetPosition(rnd.Next(10, 100), rnd.Next(10, 100), pictureBoxCars.Width,
+           pictureBoxCars.Height);
+            Draw();
+        }
     }
 }
