+++ Maleev_V_A_ISEbd21/Itest.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Maleev_V_A_ISEbd21
+{
+    interface Itest
+    {
+        void SetPosition(int x, int y, int width, int height);
+        /// <summary>
+        /// Изменение направления пермещения
+        /// </summary>
+        /// <param name="direction">Направление</param>
+        void MoveTransport(Direction direction);
+        /// <summary>
+        /// Отрисовка автомобиля
+        /// </summary>
+        /// <param name="g"></param>
+        void DrawTruck(Graphics g);
+    }
+}
