+++ Maleev_V_A_ISEbd21/Vehicle.cs
@@ -1,0 +1,48 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Maleev_V_A_ISEbd21
+{
+    public abstract class Vehicle : Itest
+    {
+        /// <summary>
+        /// Левая координата отрисовки автомобиля
+        /// </summary>
+        protected float _startPosX;
+        /// <summary>
+        /// Правая кооридната отрисовки автомобиля
+        /// </summary>
+        protected float _startPosY;
+        /// <summary>
+        /// Ширина окна отрисовки
+        /// </summary>
+        protected int _pictureWidth;
+        //Высота окна отрисовки
+        protected int _pictureHeight;
+        /// <summary>
+        /// Максимальная скорость
+        /// </summary>
+        public int MaxSpeed { protected set; get; }
+        /// <summary>
+        /// Вес автомобиля
+        /// </summary>
+        public float Weight { protected set; get; }
+        /// <summary>
+        /// Основной цвет кузова
+        /// </summary>
+        public Color MainColor { protected set; get; }
+        public void SetPosition(int x, int y, int width, int height)
+        {
+            _startPosX = x;
+            _startPosY = y;
+            _pictureWidth = width;
+            _pictureHeight = height;
+        }
+        public abstract void DrawTruck(Graphics g);
+        public abstract void MoveTransport(Direction direction);
+    }
+}
